+++ src/main/java/com/kard/api/resources/transactions/AsyncRawTransactionsClient.java
@@ -494,7 +494,7 @@ public class AsyncRawTransactionsClient {
     }
 
     /**
-     * Retrieve rewarded transaction history for a specific user. Returns only SETTLED transactions within the last 12 months.
+     * Retrieve rewarded transaction history for a specific user. By default this returns only SETTLED transactions within the last 12 months.
      * <br/>
      * <b>Required scopes:</b> <code>transaction:read</code>
      * <br/>
@@ -507,7 +507,7 @@ public class AsyncRawTransactionsClient {
     }
 
     /**
-     * Retrieve rewarded transaction history for a specific user. Returns only SETTLED transactions within the last 12 months.
+     * Retrieve rewarded transaction history for a specific user. By default this returns only SETTLED transactions within the last 12 months.
      * <br/>
      * <b>Required scopes:</b> <code>transaction:read</code>
      * <br/>
@@ -520,7 +520,7 @@ public class AsyncRawTransactionsClient {
     }
 
     /**
-     * Retrieve rewarded transaction history for a specific user. Returns only SETTLED transactions within the last 12 months.
+     * Retrieve rewarded transaction history for a specific user. By default this returns only SETTLED transactions within the last 12 months.
      * <br/>
      * <b>Required scopes:</b> <code>transaction:read</code>
      * <br/>
@@ -532,7 +532,7 @@ public class AsyncRawTransactionsClient {
     }
 
     /**
-     * Retrieve rewarded transaction history for a specific user. Returns only SETTLED transactions within the last 12 months.
+     * Retrieve rewarded transaction history for a specific user. By default this returns only SETTLED transactions within the last 12 months.
      * <br/>
      * <b>Required scopes:</b> <code>transaction:read</code>
      * <br/>
@@ -558,6 +558,10 @@ public class AsyncRawTransactionsClient {
         if (request.getPageSize().isPresent()) {
             QueryStringMapper.addQueryParameter(
                     httpUrl, "page[size]", request.getPageSize().get(), false);
+        }
+        if (request.getFilterStatus().isPresent()) {
+            QueryStringMapper.addQueryParameter(
+                    httpUrl, "filter[status]", request.getFilterStatus().get(), false);
         }
         if (request.getInclude().isPresent()) {
             QueryStringMapper.addQueryParameter(